removed unused imports, replaced some internal JDK exceptions and added the equivalents from apache.commons-lang, added some try-with-resources

+++ alien4cloud-common/src/main/java/alien4cloud/utils/ImageResizeUtil.java
@@ -90,8 +90,8 @@ public final class ImageResizeUtil {
         int targetWidth = width;
         int targetHeight = height;
 
-        float targetDimensions = Float.valueOf(width).floatValue() / Float.valueOf(height).floatValue();
-        float sourceDimensions = Float.valueOf(originalWidth).floatValue() / Float.valueOf(originalHeight).floatValue();
+        float targetDimensions = (float) width / (float) height;
+        float sourceDimensions = (float) originalWidth / (float) originalHeight;
         if (targetDimensions > sourceDimensions) {
             targetWidth = Float.valueOf(width * sourceDimensions / targetDimensions).intValue();
         } else {
